move timetable java files
