+++ app/src/main/java/sg/edu/np/mad/inkwell/TimetableActivity.java
@@ -1,4 +1,4 @@
-package sg.edu.np.mad.inkwell.ui.timetable;
+package sg.edu.np.mad.inkwell;
 
 import android.app.TimePickerDialog;
 import android.os.Bundle;
@@ -11,24 +11,17 @@ import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.TimePicker;
 
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.cardview.widget.CardView;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import java.sql.Time;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 import java.util.Locale;
-
-import sg.edu.np.mad.inkwell.R;
 
 public class TimetableActivity extends AppCompatActivity {
